US101 - feat(auth): implement company registration with admin user
- added endpoint to register company and admin user
- truncated passwords to 72 chars for bcrypt
- updated models: autoincrement IDs, hashed_password String length
- added integration tests for registration flow

+++ app/db/models/user.py
@@ -1,11 +1,10 @@
 # app/db/models/user.py
 from __future__ import annotations
 
-import uuid
 from enum import Enum as PyEnum
 from typing import TYPE_CHECKING
 
-from sqlalchemy import BigInteger, ForeignKey, String, Enum as SAEnum
+from sqlalchemy import ForeignKey, Integer, String, Enum as SAEnum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 
@@ -21,7 +20,7 @@ class Role(PyEnum):
 class User(Base):
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(50), nullable=False)
     username: Mapped[str] = mapped_column(String(50),nullable=False, unique=True)
     hashed_password: Mapped[str] = mapped_column(String(255), nullable=False)
@@ -31,12 +30,12 @@ class User(Base):
         default=Role.EMPLOYEE,
     )
     company_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("companies.id"),
         nullable=False
     )
     branch_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("branches.id"),
         nullable=True
     )
